check that tex and bib files exist in directory

+++ turnPdf.sh
@@ -6,30 +6,34 @@ echo "Please ensure that the .tex and .bib file have the same file name, ignorin
 echo -n -e "\nWhat is the filename? "
 read filename
 
-echo -e "\nCompiling .tex file..."
-pdflatex $filename 1> /dev/null
-
-echo -e "\nCompiling .bib file..."
-biber $filename 1> /dev/null
-
-echo -e "\nCompiling .tex file again..."
-pdflatex $filename 1> /dev/null
-
-
 texName=$filename".tex"
 bibName=$filename".bib"
 pdfName=$filename".pdf"
 
-for file in *; do 
-    if [ -f "$file" ]; then 
-        if [[ "$file" = "$texName" || "$file" = "$bibName" || "$file" = "$pdfName" || "$file" = "turnPdf.sh" ]]; then
-            echo -e "\nDeleting unneeded files..."
-        else 
-            rm "$file" 
-        fi
-    fi 
-done
+if [[ -f "$texName" || -f "$bibName" ]]; then 
 
-echo -e "\nYou now have the corresponding PDF."
-echo -e "\nHere are the stats for your document:"
-texcount $filename.tex
+    echo -e "\nCompiling .tex file..."
+    pdflatex $filename 1> /dev/null
+
+    echo -e "\nCompiling .bib file..."
+    biber $filename 1> /dev/null
+
+    echo -e "\nCompiling .tex file again..."
+    pdflatex $filename 1> /dev/null
+
+    for file in *; do 
+        if [ -f "$file" ]; then 
+            if [[ "$file" = "$texName" || "$file" = "$bibName" || "$file" = "$pdfName" || "$file" = "turnPdf.sh" ]]; then
+                echo -e "\nDeleting unneeded files..."
+            else 
+                rm "$file" 
+            fi
+        fi 
+    done
+
+    echo -e "\nYou now have the corresponding PDF."
+    echo -e "\nHere are the stats for your document:"
+    texcount $filename.tex
+else
+    echo -e "\nMake sure that both files exist or have matching names.\n"
+fi
